Don't send string "null" for all config options
When you install the app, all config options are sent to the config page as the string "null". This is annoying as you have to clear them all yourself. Also it prevents skipping blank codes. It appears that the variable state should be set within the showConfiguration handler.
Tested on Android with Pebble Time and in the CloudPebble emulator.